- Moved physx lib inclusion to project - Added final libraries to additional include dirs

+++ FluidSandbox/OSLowLevel.cpp
@@ -41,7 +41,7 @@ const std::string COSLowLevel::getTextFileContent(const std::string &filename)
 const uint8_t *COSLowLevel::getBinaryFileContent(const std::string &filename) {
 	std::ifstream myfile(filename, std::ios::binary | std::ios::ate);
 	if(myfile && myfile.is_open()) {
-		size_t size = myfile.tellg();
+		size_t size = (size_t)myfile.tellg();
 		uint8_t *result = new uint8_t[size];
 		myfile.seekg(0, std::ios::beg);
 		myfile.read((char *)result, size);
